Implemented test for Bomber and adapted all the class for the test

+++ src/main/java/bomberOne/model/bomber/PowerUpHandler.java
@@ -11,7 +11,7 @@ public interface PowerUpHandler {
 	/**
 	 * This method apply the powerUp More Speed
 	 */
-	public void applySpeed(int increment);
+	public void applySpeed(double increment);
 	
 	/**
 	 * This method apply the powerUp Pierce
